refactor(UserChallenges): renaming form and item, moving form constants and types to UserChallenges
