--- src/modules/UserChallenges/UserChallenge.form.constants.ts
@@ -1,17 +1,0 @@
-import { ChallengeFields } from './UserChallenge.form.types'
-
-export const INITIAL_VALUES: ChallengeFields = {
-  challenge: '',
-  challengeHostProvider: '',
-  description: '',
-  id: '',
-  startDate: null,
-  endDate: null,
-  skillNames: [],
-}
-
-export const MOCK_SKILLS = [
-  { label: 'UI', value: 'UI' },
-  { label: 'Design', value: 'Design' },
-  { label: 'UX', value: 'UX' },
-]
