--- src/modules/UserChallenges/UserChallenge.form.types.ts
@@ -1,9 +1,0 @@
-export type ChallengeFields = {
-  challenge: string
-  challengeHostProvider: string
-  description: string
-  id: string
-  startDate: Date | null
-  endDate: Date | null
-  skillNames: string[]
-}
