+++ src/modules/UserChallenges/UserChallengeItem.styles.ts
