+++ src/modules/UserChallenges/UserChallenges.constants.ts
@@ -1,3 +1,5 @@
+import { UserChallengeFormFields } from './UserChallenges.types'
+
 export const MOCKED_CHALLENGES = [
   {
     challenge: 'COVID Challenge',
@@ -15,4 +17,20 @@ export const MOCKED_CHALLENGES = [
     endDate: '04/01/2020',
     description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ',
   },
+]
+
+export const INITIAL_VALUES: UserChallengeFormFields = {
+  challenge: '',
+  challengeHostProvider: '',
+  description: '',
+  id: '',
+  startDate: null,
+  endDate: null,
+  skillNames: [],
+}
+
+export const MOCK_SKILLS = [
+  { label: 'UI', value: 'UI' },
+  { label: 'Design', value: 'Design' },
+  { label: 'UX', value: 'UX' },
 ]
